16 - Updated functionality so that out event sets of loop end events indicate a branch if one exists in the data. Changed location of functions and tests due to circular imports
